AUTO MONTAGE/requirements.txt
google-cloud-speech==2.8.0
moviepy==1.0.3